+++ addons/confirma/src/attributes/CategoryAttribute.cs
@@ -2,10 +2,10 @@ using System;
 
 namespace Confirma.Attributes;
 
-[AttributeUsage(AttributeTargets.Method)]
+[AttributeUsage(AttributeTargets.Class, Inherited = false, AllowMultiple = true)]
 public class CategoryAttribute : Attribute
 {
-    public string Category { get; }
+    public string Category { get; init; }
 
     public CategoryAttribute(string category)
     {